refactor #6 메인 페이지 타입 및 mock 구조 정리
