+++ src/pages/main/mocks/mockRooms.ts
@@ -1,14 +1,4 @@
-export interface MainRoom {
-  id: number;
-  title: string;
-  roomType: 'MIXED' | 'FEMALE' | 'MALE';
-  minAge: number;
-  maxAge: number;
-  currentCount: number;
-  capacity: number;
-  place: string;
-  lunchAt: string;
-}
+import type { MainRoom } from '../model/types';
 
 export const mockRooms: MainRoom[] = [
   {
