+++ src/pages/main/model/types.ts
@@ -1,0 +1,11 @@
+export interface MainRoom {
+  id: number;
+  title: string;
+  roomType: 'MIXED' | 'FEMALE' | 'MALE';
+  minAge: number;
+  maxAge: number;
+  currentCount: number;
+  capacity: number;
+  place: string;
+  lunchAt: string;
+}
